harness/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prompt-optimization-harness"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "dspy-ai>=3.1.0",
    "anthropic>=0.28.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]
